move ClassAccessVisitor to inspection class
